Are you sure promptovi svugdje

+++ src/main/java/production/bussines_contacts/partials/DeletableCell.java
@@ -5,6 +5,8 @@ import javafx.scene.control.TableCell;
 import javafx.scene.control.TableView;
 import production.bussines_contacts.interfaces.Deletable; // You need to create this interface
 
+import static production.bussines_contacts.utils.FunctionUtils.confirmDeleteOperation;
+
 public class DeletableCell<T extends Deletable> extends TableCell<T, Void> {
     private final Button deleteButton;
 
@@ -16,6 +18,9 @@ public class DeletableCell<T extends Deletable> extends TableCell<T, Void> {
     private void handleDeleteAction() {
         int index = getIndex();
         T item = getTableView().getItems().get(index);
+        if(!confirmDeleteOperation(item.deleteText())) {
+            return;
+        }
         if (item != null) {
             item.delete(); // Assuming T has a delete method defined in Deletable interface
             getTableView().getItems().remove(item); // Or refresh the table view as needed
